Fix integer type warnings in z85 divide by 85 macro
The two warnings were generated while compiling on a 32bit ARM, they were:

warning: this decimal constant is unsigned only in ISO C90 [enabled by default]
warning: right shift count >= width of type [enabled by default]

They both originated from the fact that the 'number' in the macro was being
multiplied by a 64bit constant. However, the constant as defined was being
inferred as a signed 32bit int upon creation. This was solved by using the
'ULL' (unsigned long long) suffix with the value. A different fix might be to
use the UINT64_C macro which is defined in 'stdint.h'. But this would require
the removal of the custom uint32_t etc defines.